[ADD] WIP code_generator_model_builder    [ADD] Refactor code_generator and format all code [FIX] remove model inter-dependency magic number [ADD] code_generator: support more mysql type [ADD] code_generator_db_servers columns views when load db - show exception when failure load db [UPD] code_generator_db_servers: format code [ADD] code_generator_db_servers: support primary_key in fields [FIX] icon with white background [FIX] database typo and move configuration at the end of menu [ADD] code_generator_db_servers with more model type [ADD] code_generator_db_servers requirements.txt [ADD] code_generator_db_servers add tree views, menu, access and action [FIX] add view generation record, menu and tree views - Order record menu by id creation order, the root need to be first - Change how create menues [FIX] remove wrong help in actions [FIX] code_generator unlink o2m_models [FIX] code_generator move generate views in model_builder [ADD] code_generator_model_builder: wizard clear views and select model [ADD] wip code_generator_generate_portal_wizard [FIX] typo menues to menus [FIX] typo BLANCK to BLANK [FIX] code_generator: comment lines cause crash when using pydev - this line is used to validate field rec_name when it do not exist - comment this will cause bug of wrong use rec_name later in execution [FIX] code_generator: hide option copy=False [ADD] code_generator_model_builder: fix clear menus [FIX] support icon when download module - hide action set icon because not working - hack force icon until the last will be fix [FIX] code_generator: remove jinja parse because conflict with odoo ways - cannot execute function doesn't exist when create the module, it tries to transform python code [FIX] rename typo serverconstrains to server_constrains - clean code [ADD] code_generator: feature to sync with local code with generation code [FIX] french in ir.model.access.csv and remove unused xml version tag Remove header coding utf-8 [FIX] blacklist and whitelist fields for export data

+++ code_generator/models/code_generator.py
@@ -99,6 +99,18 @@ class CodeGeneratorModule(models.Model):
         'm2o_module'
     )
 
+    o2m_nomenclator_whitelist_fields = fields.One2many(
+        'code.generator.ir.model.fields',
+        'm2o_module',
+        domain=[("nomenclature_whitelist", "=", True)]
+    )
+
+    o2m_nomenclator_blacklist_fields = fields.One2many(
+        'code.generator.ir.model.fields',
+        'm2o_module',
+        domain=[("nomenclature_blacklist", "=", True)]
+    )
+
     o2m_model_access = fields.One2many(
         'ir.model.access',
         compute='_get_models_info'
@@ -119,6 +131,16 @@ class CodeGeneratorModule(models.Model):
         compute='_get_models_info'
     )
 
+    o2m_model_act_url = fields.One2many(
+        comodel_name='ir.actions.act_url',
+        inverse_name='m2o_code_generator'
+    )
+
+    o2m_model_act_todo = fields.One2many(
+        comodel_name='ir.actions.todo',
+        inverse_name='m2o_code_generator'
+    )
+
     o2m_model_act_window = fields.One2many(
         'ir.actions.act_window',
         compute='_get_models_info'
@@ -129,7 +151,7 @@ class CodeGeneratorModule(models.Model):
         compute='_get_models_info'
     )
 
-    o2m_model_serverconstrains = fields.One2many(
+    o2m_model_server_constrains = fields.One2many(
         'ir.model.server_constrain',
         compute='_get_models_info'
     )
@@ -138,6 +160,24 @@ class CodeGeneratorModule(models.Model):
         'ir.actions.report',
         compute='_get_models_info'
     )
+
+    o2m_menus = fields.One2many(
+        'ir.ui.menu',
+        'm2o_module',
+        context={'ir.ui.menu.full_list': True}
+    )
+
+    nomenclator_only = fields.Boolean(string="Only export data", default=False,
+                                      help="Useful to export data with existing model.")
+
+    # Dev binding code
+    enable_sync_code = fields.Boolean(string="Enable Sync Code", default=False,
+                                      help="Will sync with code on drive when generate.")
+
+    # TODO default path actual path of this file
+    path_sync_code = fields.Char(string="Path", help="Path where sync the code.")
+
+    # clean_before_sync_code = fields.Boolean(string="Clean before sync", help="Clean before sync, all will be lost.")
 
     @api.depends('o2m_models')
     def _get_models_info(self):
@@ -148,14 +188,8 @@ class CodeGeneratorModule(models.Model):
             module.o2m_model_views = module.o2m_models.mapped('view_ids')
             module.o2m_model_act_window = module.o2m_models.mapped('o2m_act_window')
             module.o2m_model_act_server = module.o2m_models.mapped('o2m_server_action')
-            module.o2m_model_serverconstrains = module.o2m_models.mapped('o2m_serverconstrains')
+            module.o2m_model_server_constrains = module.o2m_models.mapped('o2m_server_constrains')
             module.o2m_model_reports = module.o2m_models.mapped('o2m_reports')
-
-    o2m_menus = fields.One2many(
-        'ir.ui.menu',
-        'm2o_module',
-        context={'ir.ui.menu.full_list': True}
-    )
 
     @api.depends('name', 'description')
     def _get_desc(self):
@@ -206,8 +240,9 @@ class CodeGeneratorModule(models.Model):
     @api.multi
     def unlink(self):
         o2m_models = self.mapped('o2m_models')
-        o2m_models.mapped('view_ids').unlink()
-        o2m_models.unlink()  # I need to delete the ceated tables
+        if o2m_models:
+            o2m_models.mapped('view_ids').unlink()
+            o2m_models.unlink()  # I need to delete the created tables
         return super(CodeGeneratorModule, self).unlink()
